Ran 2to3 on all .py files

+++ theano/misc/cudamat_utils.py
@@ -95,7 +95,7 @@ try:
             strides = tuple(strides[::-1])
 
             import ctypes
-            ptr_long = long(ctypes.cast(x.mat.data_device, ctypes.c_void_p).value)
+            ptr_long = int(ctypes.cast(x.mat.data_device, ctypes.c_void_p).value)
 
             # seems legit.
             z = cuda.from_gpu_pointer(ptr_long, x.shape, strides, x)
