+++ theano/misc/gnumpy_utils.py
@@ -2,6 +2,7 @@
 This code can only work if gnumpy and theano are initialized on the
 same gpu as theano.
 """
+from functools import reduce
 
 try:
     import gnumpy
@@ -107,7 +108,7 @@ try:
             strides = tuple(strides)
 
             import ctypes
-            ptr_long = long(ctypes.cast(x._base.mat.data_device, ctypes.c_void_p).value)
+            ptr_long = int(ctypes.cast(x._base.mat.data_device, ctypes.c_void_p).value)
 
             # seems legit.
             z = cuda.from_gpu_pointer(ptr_long, x.shape, strides, x._base)
